rework a lot off our bitfoo
New ideas from the insomniac optimization front, after a bliss
insight what is so cool about altivec

+++ G2QHT.c
@@ -394,7 +394,7 @@ void g2_qht_frag_clean(struct qht_fragment *to_clean)
 	memset(to_clean, 0, sizeof(*to_clean));
 }
 
-uint32_t g2_qht_global_get_ent(void)
+size_t g2_qht_global_get_ent(void)
 {
 	return global_qht->entries;
 }
